\name{aitchison360}
\alias{aitchison360}
\docType{data}
\title{ Another data set from the Atichsion book }
\description{
  AFM compositions of 23 aphyric Skye lavas. This data set can be found on page 360 of the Aitchison book (see reference).
}
\usage{data(aitchison360)}
\format{
  A data frame with 23 observations on the following 3 variables.
  \describe{
    \item{\code{V1}}{a numeric vector of percentages of Na2O\eqn{+}K2O}
    \item{\code{V2}}{a numeric vector of percentages of Fe2O3}
    \item{\code{V3}}{a numeric vector of percentages of MgO}
  }
}
\details{
}
\source{
Aitchison, J. (1986) \emph{The Statistical Analysis of Compositional
Data} Monographs on Statistics and Applied Probability. Chapman \&
Hall Ltd., London (UK). 416p.
}
\examples{
data(aitchison360)
}
\keyword{datasets}
