% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSets.R
\docType{data}
\name{alcohol}
\alias{alcohol}
\title{alcohol consumptions by country and type of alcohol}
\format{
A data frame with 193 rows and 6 variables
}
\source{
Transfered from the World Health Organisation website.
}
\usage{
data(alcohol)
}
\description{
\itemize{
 \item{\code{country }}{Country}
 \item{\code{year }}{Year}
 \item{\code{beer }}{Consumption of pure alcohol on beer (in percentages)}
 \item{\code{wine }}{Consumption of pure alcohol on wine (in percentages)}
 \item{\code{spirits }}{Consumption of pure alcohol on spirits (in percentages)}
 \item{\code{other }}{Consumption of pure alcohol on other beverages (in percentages)}
}
}
\examples{

data("alcohol")
str(alcohol)
summary(alcohol)
}
\author{
Matthias Templ \email{matthias.templ@tuwien.ac.at}
}
\keyword{data}
