% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balZav.R
\name{balZav}
\alias{balZav}
\title{New average symmetric coordinates}
\usage{
balZav(x)
}
\arguments{
\item{x}{compositional data}
}
\value{
\item{Z.av}{A matrix of new average symmetric coordinates.}
}
\description{
New average symmetric coordinates
}
\author{
Petra Kynclova
}
