% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dataSets.R
\docType{data}
\name{cancer}
\alias{cancer}
\title{Hospital discharges on cancer and distribution of age}
\format{A data set on 24 compositions on 6 variables.}
\source{
\url{http://www.ec.europa.eu/eurostat} and \url{http://unstats.un.org/unsd}
}
\description{
Hospital discharges of in-patients on neoplasms (cancer) per 100.000 inhabitants (year 2007) and population age structure.
}
\details{
\itemize{
\item{\code{country }}{country}
\item{\code{year }}{year}
\item{\code{<15 }}{percentage of population with age below 15}
\item{\code{15-60 }}{percentage of population with age between 15 and 60}
\item{\code{60+ }}{percentage of population with age above 60}
\item{\code{discharges }}{hospital discharges of in-patients on neoplasms (cancer) per 100.000 inhabitants}
}

The response (discharges) is provided for the European Union countries (except Greece, Hungary and Malta) by Eurostat. As explanatory variables we use the age structure of the population in the same countries (year 2008). The age structure consists of three parts, age <15, age 15–60 and age >60 years, and they are expressed as percentages on the overall population in the countries. The data are provided by the United Nations Statistics Division.
}
\examples{
data(cancer)
str(cancer)
}
\author{
conversion to R by Karel Hron and Matthias Templ \email{matthias.templ@tuwien.ac.at}
}
\references{
K. Hron, P. Filzmoser, K. Thompson (2012). Linear regression with compositional explanatory variables. \emph{Journal of Applied Statistics}, Volume 39, Issue 5, 2012.
}
\keyword{datasets}

