% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSets.R
\docType{data}
\name{cancerMN}
\alias{cancerMN}
\title{malignant neoplasms cancer}
\format{
A data set on 35 compositional tables on 4 parts (row-wise sorted cells) and 5 variables.
}
\source{
From OECD website
}
\description{
Two main types of malignant neoplasms cancer affecting colon and lung, respectively, in male and female populations. 
For this purpose population data (2012) from 35 OECD countries were collected.
}
\details{
\itemize{
\item{\code{country }}{country}                                  
\item{\code{females-colon }}{number of colon cancer cases in female population}                       
\item{\code{females-lung }}{number of lung cancer cases in female population}                    
\item{\code{males-colon }}{number of colon cancer cases in male population}
\item{\code{males-lung }}{number of lung cancer cases in male population}
}

The data are obtained from the OECD website.
}
\examples{

data(cancerMN)
head(cancerMN)
rowSums(cancerMN[, 2:5])
}
\author{
conversion to R by Karel Hron and intergration by Matthias Templ \email{matthias.templ@tuwien.ac.at}
}
\keyword{datasets}
