% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ced.R
\name{ced}
\alias{ced}
\title{Compositional error deviation}
\usage{
ced(x, y, ni)
}
\arguments{
\item{x}{matrix or data frame}

\item{y}{matrix or data frame of the same size as x}

\item{ni}{normalization parameter. See details below.}
}
\value{
the compositinal error distance
}
\description{
Normalized Aitchison distance between two data sets
}
\details{
This function has been mainly written for procudures 
that evaluate imputation or replacement of rounded zeros. The ni parameter can thus, e.g. be
used for expressing the number of rounded zeros.
}
\examples{
data(expenditures)
x <- expenditures
x[1,3] <- NA
xi <- impKNNa(x)$xImp
ced(expenditures, xi, ni = sum(is.na(x)))
}
\references{
Hron, K., Templ, M., Filzmoser, P. (2010) Imputation of
missing values for compositional data using classical and robust methods
\emph{Computational Statistics and Data Analysis}, 54 (12),
3095-3107.

Templ, M., Hron, K., Filzmoser, P., Gardlo, A. (2016). 
Imputation of rounded zeros for high-dimensional compositional data. 
\emph{Chemometrics and Intelligent Laboratory Systems}, 155, 183-190.
}
\seealso{
\code{\link{rdcm}}
}
\author{
Matthias Templ
}
\keyword{manip}
