% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoDaCuCoord.R
\name{CoDaCuCoord}
\alias{CoDaCuCoord}
\title{Coordinate representation of a compositional cube}
\usage{
CoDaCuCoord(x, row.factor = NULL, col.factor = NULL,
  slice.factor = NULL, value = NULL, SBPr = NULL, SBPc = NULL,
  SBPs = NULL, pivot = FALSE, print.res = FALSE)
}
\arguments{
\item{x}{a data frame containing variables representing row, column and slice factors of the respective compositional cube and variable with the values of the composition.}

\item{row.factor}{name of the variable representing the row factor. Needs to be stated with the quotation marks.}

\item{col.factor}{name of the variable representing the column factor. Needs to be stated with the quotation marks.}

\item{slice.factor}{name of the variable representing the slice factor. Needs to be stated with the quotation marks.}

\item{value}{name of the variable representing the values of the composition. Needs to be stated with the quotation marks.}

\item{SBPr}{an \eqn{I-1\times I} array defining the sequential binary partition of the values of the row factor, where I is the number of the row factor levels. The values assigned in the given step to the + group are marked by 1, values from  the - group by -1 and the rest by 0. If it is not provided, the pivot version of coordinates is constructed automatically.}

\item{SBPc}{an \eqn{J-1\times J} array defining the sequential binary partition of the values of the column factor, where J is the number of the column factor levels. The values assigned in the given step to the + group are marked by 1, values from  the - group by -1 and the rest by 0. If it is not provided, the pivot version of coordinates is constructed automatically.}

\item{SBPs}{an \eqn{K-1\times K} array defining the sequential binary partition of the values of the slice factor, where K is the number of the slice factor levels. The values assigned in the given step to the + group are marked by 1, values from  the - group by -1 and the rest by 0. If it is not provided, the pivot version of coordinates is constructed automatically.}

\item{pivot}{logical, default is FALSE. If TRUE, or one of the SBPs is not defined, its pivot version is used.}

\item{print.res}{logical, default is FALSE. If TRUE, the output is displayed in the Console.}
}
\value{
\item{Coordinates}{an array of orthonormal coordinates.} 
\item{Grap.rep}{graphical representation of the coordinates. 
Parts denoted by + form the groups in the numerator of the respective computational formula, 
parts - form the denominator and parts . are not involved in the given coordinate.} 
\item{Row.balances}{an array of row balances.}
\item{Column.balances}{an array of column balances.}
\item{Slice.balances}{an array of slice balances.}
\item{Row.column.OR}{an array of row-column OR coordinates.}
\item{Row.slice.OR}{an array of row-slice OR coordinates.}
\item{Column.slice.OR}{an array of column-slice OR coordinates.}
\item{Row.col.slice.OR}{an array of coordinates describing the mutual interaction between all three factors.}
\item{Contrast.matrix}{contrast matrix.}
\item{Log.ratios}{an array of pure log-ratios between groups of parts without the normalizing constant.}
\item{Coda.cube}{cube form of the given composition.}
}
\description{
CoDaCuCoord computes a system of orthonormal coordinates of a compositional cube. 
Computation of either pivot coordinates or a coordinate system based on the given SBP is possible.
}
\details{
CoDaCuCoord

This transformation moves the IJK-part compositional cubes from the simplex into a (IJK-1)-dimensional real space isometrically with respect to its three-factorial nature.
}
\examples{
### example from Fa\\v cevicov\\'a (2019)
data(employment2)
CZE <- employment2[which(employment2$Country == 'CZE'), ]

# pivot coordinates
CoDaCuCoord(CZE, "Sex", 'Contract', "Age", 'Value')

# coordinates with given SBP

r <- t(c(1,-1))
c <- t(c(1,-1))
s <- rbind(c(1,-1,-1), c(0,1,-1))

CoDaCuCoord(CZE, "Sex", 'Contract', "Age", 'Value', r,c,s)
}
\references{
Facevicova, K., Filzmoser, P. and K. Hron (2019) Compositional Cubes: Three-factorial Compositional Data. Under review.
}
\seealso{
\code{\link{CoDaTaCoord}} 
\code{\link{CoDaTaCoordWrapper}} 
\code{\link{CoDaCuCoordWrapper}}
}
\author{
Kamila Facevicova
}
\keyword{coordinates}
\keyword{multivariate,}
