% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoDaTaCoord.R
\name{CoDaTaCoord}
\alias{CoDaTaCoord}
\title{Coordinate representation of a compositional cube}
\usage{
CoDaTaCoord(x = NULL, row.factor = NULL, col.factor = NULL,
  value = NULL, SBPr = NULL, SBPc = NULL, pivot = FALSE,
  print.res = FALSE)
}
\arguments{
\item{x}{a data frame containing variables representing row and column factors of the respective compositional table and variable with the values of the composition.}

\item{row.factor}{name of the variable representing the row factor. Needs to be stated with the quotation marks.}

\item{col.factor}{name of the variable representing the column factor. Needs to be stated with the quotation marks.}

\item{value}{name of the variable representing the values of the composition. Needs to be stated with the quotation marks.}

\item{SBPr}{an \eqn{I-1\times I} array defining the sequential binary partition of the values of the row factor, where I is the number of the row factor levels. The values assigned in the given step to the + group are marked by 1, values from  the - group by -1 and the rest by 0. If it is not provided, the pivot version of coordinates is constructed automatically.}

\item{SBPc}{an \eqn{J-1\times J} array defining the sequential binary partition of the values of the column factor, where J is the number of the column factor levels. The values assigned in the given step to the + group are marked by 1, values from  the - group by -1 and the rest by 0. If it is not provided, the pivot version of coordinates is constructed automatically.}

\item{pivot}{logical, default is FALSE. If TRUE, or one of the SBPs is not defined, its pivot version is used.}

\item{print.res}{logical, default is FALSE. If TRUE, the output is displayed in the Console.}
}
\value{
\item{Coordinates}{an array of orthonormal coordinates.}
\item{Grap.rep}{graphical representation of the coordinates. Parts denoted by \verb{+} form the groups in the numerator of the respective computational formula, parts \verb{-} form the denominator and parts \verb{.} are not involved in the given coordinate.}
\item{Ind.coord}{an array of row and column balances. Coordinate representation of the independent part of the table.}
\item{Int.coord}{an array of OR coordinates. Coordinate representation of the interactive part of the table.}
\item{Contrast.matrix}{contrast matrix.}
\item{Log.ratios}{an array of pure log-ratios between groups of parts without the normalizing constant.}
\item{Coda.table}{table form of the given composition.}
}
\description{
CoDaTaCoord computes a system of orthonormal coordinates of a compositional table. 
Computation of either pivot coordinates or a coordinate system based on the given SBP is possible.
}
\details{
CoDaTaCoord

This transformation moves the IJ-part compositional tables 
from the simplex into a (IJ-1)-dimensional real space isometrically 
with respect to its two-factorial nature. 
The coordinate system is formed by two types of coordinates - balances and log odds-ratios.
}
\examples{
# example from Fa\\v cevicov\\'a (2018):
data(manu_abs)
manu_USA <- manu_abs[which(manu_abs$country=='USA'),]
manu_USA$output <- factor(manu_USA$output, levels=c('LAB', 'SUR', 'INP'))

# pivot coordinates
CoDaTaCoord(manu_USA, row.factor = 'output', col.factor = 'isic', value='value')

# SBPs defined in paper
r <- rbind(c(-1,-1,1), c(-1,1,0))
c <- rbind(c(-1,-1,-1,-1,1), c(-1,-1,-1,1,0), c(-1,-1,1,0,0), c(-1,1,0,0,0))
CoDaTaCoord(manu_USA, row.factor = 'output', col.factor = 'isic', value='value', SBPr=r, SBPc=c)
}
\references{
Facevicova, K., Hron, K., Todorov, V. and M. Templ (2018) 
General approach to coordinate representation of compositional tables. 
Scandinavian Journal of Statistics, 45(4), 879--899.
}
\seealso{
\code{\link{CoDaCuCoord}} 
\code{\link{CoDaTaCoordWrapper}} 
\code{\link{CoDaCuCoordWrapper}}
}
\author{
Kamila Facevicova
}
\keyword{coordinates}
\keyword{multivariate,}
