% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoDaTaCoordWrapper.R
\name{CoDaTaCoordWrapper}
\alias{CoDaTaCoordWrapper}
\title{Coordinate representation of a sample of compositional tables}
\usage{
CoDaTaCoordWrapper(X, obs.ID = NULL, row.factor = NULL,
  col.factor = NULL, value = NULL, SBPr = NULL, SBPc = NULL,
  pivot = FALSE, test = FALSE, n.boot = 1000)
}
\arguments{
\item{X}{a data frame containing variables representing row and column factors of the respective compositional tables, variable with the values 
of the composition and variable distinguishing the observations.}

\item{obs.ID}{name of the variable distinguishing the observations. Needs to be stated with the quotation marks.}

\item{row.factor}{name of the variable representing the row factor. Needs to be stated with the quotation marks.}

\item{col.factor}{name of the variable representing the column factor. Needs to be stated with the quotation marks.}

\item{value}{name of the variable representing the values of the composition. Needs to be stated with the quotation marks.}

\item{SBPr}{an \eqn{I-1\times I} array defining the sequential binary partition of the values of the row factor, where I is the number of the row factor levels. The values assigned in the given step to the + group are marked by 1, values from  the - group by -1 and the rest by 0. If it is not provided, the pivot version of coordinates is constructed automatically.}

\item{SBPc}{an \eqn{J-1\times J} array defining the sequential binary partition of the values of the column factor, where J is the number of the column factor levels. The values assigned in the given step to the + group are marked by 1, values from  the - group by -1 and the rest by 0. If it is not provided, the pivot version of coordinates is constructed automatically.}

\item{pivot}{logical, default is FALSE. If TRUE, or one of the SBPs is not defined, its pivot version is used.}

\item{test}{logical, default is \code{FALSE}. If \code{TRUE}, the bootstrap analysis of coordinates is provided.}

\item{n.boot}{number of bootstrap samples.}
}
\description{
For each compositional table in the sample \code{CoDaTaCoord2} 
computes a system of orthonormal coordinates and provide a simple descriptive analysis. 
Computation of either pivot coordinates or a coordinate system based on the given SBP is possible.
}
\details{
CoDaTaCoordWrapper

Each of n IJ-part compositional tables from the sample is with 
respect to its two-factorial nature isometrically transformed from the simplex 
into a (IJ-1)-dimensional real space. Sample mean values and standard deviations are 
computed and using bootstrap an estimate of 95 \% confidence interval is given.
}
\examples{
# example from Fa\\v cevicov\\'a (2018):
data(manu_abs)

### Compositional tables approach,
### analysis of the relative structure.
### An example from Facevi\\v cov\\'a (2018)

manu_abs$output <- factor(manu_abs$output, levels=c('LAB', 'SUR', 'INP'))

# pivot coordinates
CoDaTaCoordWrapper(manu_abs, obs.ID='country',
row.factor = 'output', col.factor = 'isic', value='value')

# SBPs defined in paper
r <- rbind(c(-1,-1,1), c(-1,1,0))
c <- rbind(c(-1,-1,-1,-1,1), c(-1,-1,-1,1,0), 
c(-1,-1,1,0,0), c(-1,1,0,0,0))
CoDaTaCoordWrapper(manu_abs, obs.ID='country',row.factor = 'output', 
col.factor = 'isic', value='value', SBPr=r, SBPc=c, test=TRUE)

### Classical approach,
### generalized linear mixed effect model.

\dontrun{
library(lme4)
glmer(value~output*as.factor(isic)+(1|country),data=manu_abs,family=poisson)
}
}
\references{
Facevicova, K., Hron, K., Todorov, V. and M. Templ (2018) 
General approach to coordinate representation of compositional tables. 
Scandinavian Journal of Statistics, 45(4), 879--899.
}
\seealso{
\code{\link{CoDaCuCoord}} 
\code{\link{CoDaTaCoord}} 
\code{\link{CoDaCuCoordWrapper}}
}
\author{
Kamila Facevicova
}
\keyword{coordinates}
\keyword{multivariate,}
