% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dataSets.R
\docType{data}
\name{coffee}
\alias{coffee}
\title{coffee data set}
\format{A data frame with 27 observations on the following 4 variables.}
\usage{
data(coffee)
}
\description{
27 commercially available coffee samples of different origins.
}
\details{
\itemize{
\item{\code{metpyr }}{hydroxy-2-propanone}
\item{\code{met5 }}{methylpyrazine}
\item{\code{furfu }}{methylfurfural}
\item{\code{sort }}{a character vector}
}

In the original data set, 15 volatile compounds (descriptors of coffee aroma) were selected for a statistical analysis. We selected only three compounds (compositional parts) hydroxy-2-propanone, methylpyrazine and methylfurfural to allow for a visualization in a ternary diagram.
}
\examples{
data(coffee)
str(coffee)
summary(coffee)
}
\author{
Matthias Templ \email{matthias.templ@tuwien.ac.at}, Karel Hron
}
\references{
M. Korhonov\'a, K. Hron, D. Klimc\'ikov\'a, L. Muller, P. Bedn\'ar, and P. Bart\'ak (2009). Coffee aroma - statistical analysis of compositional data. \emph{Talanta}, 80(2): 710--715.
}
\keyword{data}

