% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompositionalSpline.R
\name{compositionalSpline}
\alias{compositionalSpline}
\title{compositional spline}
\usage{
compositionalSpline(knots, t, clrf, w, order, der, alpha, plot = FALSE)
}
\arguments{
\item{knots}{sequence of knots}

\item{t}{class midpoints}

\item{clrf}{clr transformed values at class midpoints, i.e., clr(f(t))}

\item{w}{weights}

\item{order}{order of the spline (i.e., degree + 1)}

\item{der}{lth derivation}

\item{alpha}{smoothing parameter}

\item{plot}{if TRUE, thre resulting spline is plotted}
}
\value{
\item{\code{J}}{value of the functional J}
\item{\code{ZB_coef}}{ZB-spline basis coeffcients}
\item{\code{CV}}{score of cross-validation}
\item{\code{GCV}}{score of generalized cross-validation}
}
\description{
This code implements the compositional smooting splines grounded on the theory of 
Bayes spaces.
}
\details{
Augmented knot sequence is obtained from the original knots by adding #(order-1) multiple endpoints.
}
\examples{
SepalLengthCm <- iris$Sepal.Length
Species <- iris$Species

iris1 <- SepalLengthCm[iris$Species==levels(iris$Species)[1]]
h1 <- hist(iris1, nclass = 12, plot = FALSE)

midx1 <- h1$mids
midy1 <- matrix(h1$density, nrow=1, ncol = length(h1$density), byrow=TRUE)
knots <- 7
\dontrun{
sol1 <- compositionalSpline(knots = 3, t = 2, clrf = , w =, order =, der = , alpha =)
}
}
\references{
Machalova, J., Talska, R., Hron, K. Gaba, A. Compositional splines for representation of density functions. \emph{Comput Stat} (2020). https://doi.org/10.1007/s00180-020-01042-7
}
\author{
J. Machalova \email{jitka.machalova@upol.cz}, R. Talska \email{talskarenata@seznam.cz}, K. Hron, A. Gaba
}
