% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corCoDa.R
\name{corCoDa}
\alias{corCoDa}
\title{Correlations for compositional data}
\usage{
corCoDa(x, ...)
}
\arguments{
\item{x}{a matrix or data frame with compositional data}

\item{...}{additional arguments for the function \code{\link{cor}}}
}
\value{
A compositional correlation matrix.
}
\description{
This function computes correlation coefficients between compositional parts based
on symmetric pivot coordinates.
}
\examples{
data(expenditures)
corCoDa(expenditures)
x <- arcticLake 
corCoDa(x)

}
\references{
Kynclova, P., Hron, K., Filzmoser, P. (2017)
Correlation between compositional parts based on symmetric balances.
\emph{Mathematical Geosciences}, 49(6), 777-796.
}
\author{
Petra Kynclova
}
