% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSets.R
\docType{data}
\name{economy}
\alias{economy}
\title{Economic indicators}
\format{A data frame with 30 observations and 7 variables}
\usage{
data(economy)
}
\description{
Household and government consumptions, gross captial formation and import and exports of goods and services.
}
\details{
\itemize{
 \item{\code{country }}{country name}
 \item{\code{country2 }}{country name, short version}
 \item{\code{HHconsumption }}{Household and NPISH final consumption expenditure}
 \item{\code{GOVconsumption }}{Final consumption expenditure of general government}
 \item{\code{capital }}{Gross capital formation}
 \item{\code{exports }}{Exports of goods and services}
 \item{\code{imports }}{Imports of goods and services}
}
}
\examples{

data(economy)
str(economy)
}
\author{
Peter Filzmoser, Matthias Templ \email{matthias.templ@tuwien.ac.at}
}
\references{
Eurostat, \url{http://ec.europa.eu/eurostat/data}
}
\keyword{data}

