% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSets.R
\docType{data}
\name{educFM}
\alias{educFM}
\title{education level of father (F) and mother (M)}
\format{
A data frame with 31 observations and 8 variables
}
\source{
from Eurostat,\url{http://ec.europa.eu/eurostat/}
}
\usage{
data(educFM)
}
\description{
Education level of father (F) and mother (M) in percentages of low
(l), medium (m), and high (h) of 31 countries in Europe.
}
\details{
\itemize{
 \item{\code{country }}{community code}
 \item{\code{F.l }}{percentage of females with low edcuation level}
 \item{\code{F.m }}{percentage of females with medium edcuation level}
 \item{\code{F.h }}{percentage of females with high edcuation level}
 \item{\code{F.l }}{percentage of males with low edcuation level}
 \item{\code{F.m }}{percentage of males with medium edcuation level}
 \item{\code{F.h }}{percentage of males with high edcuation level}
}
}
\examples{

data(educFM)
str(educFM)
}
\author{
Peter Filzmoser, Matthias Templ
}
\keyword{data}
