% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSets.R
\docType{data}
\name{efsa}
\alias{efsa}
\title{efsa nutrition consumption}
\format{
A data frame with 87 observations on the following 22 variables.
\itemize{ 
\item{\code{Country }}{country name}
\item{\code{Pop.Class }}{population class}
\item{\code{grains }}{Grains and grain-based products}
\item{\code{vegetables }}{Vegetables and vegetable products (including fungi)}
\item{\code{roots }}{Starchy roots and tubers}
\item{\code{nuts }}{Legumes, nuts and oilseeds}
\item{\code{fruit }}{Fruit and fruit products}
\item{\code{meat }}{Meat and meat products (including edible offal)} 
\item{\code{fish }}{Fish and other seafood (including amphibians, rept)} 
\item{\code{milk }}{Milk and dairy products}
\item{\code{eggs }}{Eggs and egg products}
\item{\code{sugar }}{Sugar and confectionary} 
\item{\code{fat }}{Animal and vegetable fats and oils} 
\item{\code{juices }}{Fruit and vegetable juice} 
\item{\code{nonalcoholic }}{Non-alcoholic beverages (excepting milk based beverages)} 
\item{\code{alcoholic }}{Alcoholic beverages} 
\item{\code{water }}{Drinking water (water without any additives)} 
\item{\code{herbs }}{Herbs, spices and condiments} 
\item{\code{small_children_food }}{Food for infants and small children} 
\item{\code{special }}{Products for special nutritional use} 
\item{\code{composite }}{Composite food (including frozen products)} 
\item{\code{snacks }}{Snacks, desserts, and other foods} 
}
}
\source{
efsa
}
\description{
Comprehensive European Food Consumption Database
}
\details{
The Comprehensive Food Consumption Database is a source of information on 
food consumption across the European Union (EU). 
The food consumption are reported in grams per day (g/day).
}
\examples{

data(efsa)

}
\keyword{datasets}
