% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSets.R
\docType{data}
\name{electionATbp}
\alias{electionATbp}
\title{Votes at presidental elections}
\format{A data frame with 16 observations and 8 variables}
\source{
\url{https://www.data.gv.at/}
}
\usage{
data(electionATbp)
}
\description{
Votes from the final election for the president in Austria 2016
}
\details{
Votes for Hofer and Van der Bellen for each district in Austria.

\itemize{
 \item{\code{GKZ }}{community code}
 \item{\code{Name }}{name of the region}
 \item{\code{Eligible }}{eligible votes}
 \item{\code{Votes_total }}{total number of votes}
 \item{\code{Votes_invalid }}{number of invalid votes}
 \item{\code{Votes_valid }}{number of valid votes}
 \item{\code{Hofer_total }}{total number of votes for Hofer}
 \item{\code{Hofer_perc }}{percentage of votes for hofer}
}
}
\examples{

data(electionATbp)
# data still have not been made tidy (to-do):
str(electionATbp)
}
\author{
Peter Filzmoser
}
\keyword{data}

