% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSets.R
\docType{data}
\name{employment_df}
\alias{employment_df}
\title{Employment in different countries by gender and status.}
\format{
A data.frame with 132 rows and 4 columns.
}
\usage{
data(employment_df)
}
\description{
\itemize{
 \item{\code{gender}}{factor}
 \item{\code{status}}{factor, defining if part or full time work}
 \item{\code{country}}{country}
 \item{\code{value}}{employment}
}
}
\examples{
data(employment_df)
head(employment_df)
}
\keyword{data}
