% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls_pb.R
\name{fBPUpChi_PLS}
\alias{fBPUpChi_PLS}
\title{Title}
\usage{
fBPUpChi_PLS(Yp, r3, b, version = "cov")
}
\arguments{
\item{Yp}{a matrix of raw compositional data with "n" rows and "D" columns/components}

\item{r3}{a response variable; can be continuous (PLS regression) or binary (PLS-DA)}

\item{b}{a given balance constructed during the procedure (contains some zero value(s))}

\item{version}{a parameter determining whether the balances are ordered according to max. covariance (default) or max. correlation}
}
\value{
A list with the following components:
\describe{
  \item{\code{bal}}{}
  \item{\code{varbal}}{}
}
}
\description{
Title
}
