% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transformations.R
\name{fcenLR}
\alias{fcenLR}
\title{fcenLR transformation (functional)}
\usage{
fcenLR(z, z_step, density)
}
\arguments{
\item{z}{grid of point defining the abscissa}

\item{z_step}{step of the grid of the abscissa}

\item{density}{grid evaluation of the lambda-density}
}
\value{
\item{\code{out}}{grid evaluation of the lambda-density in L^2(lambda)}
}
\description{
fcenLR[lambda] transformation: mapping from B^2(lambda) into L^2(lambda)
}
\examples{
# Example (normal density)
t = seq(-4.7,4.7, length = 1000)
t_step = diff(t[1:2])

mean = 0; sd = 1.5
dens.true = dnorm(t, mean, sd)

plot(t,dens.true, type="l",las=1, ylab="fcenLR(density)", 
  cex.lab=1.2,cex.axis=1.2, col="darkblue",lwd=2)
abline(h=0, col="red")

plot(t,fcenLRinv(t.fine,t_step,dens.true), type="l",las=1, 
  ylab="density",cex.lab=1.2,cex.axis=1.2, col="darkblue",lwd=2)
}
\references{
Talska, R., Menafoglio, A., Hron, K., Egozcue, J. J., Palarea-Albaladejo, J. (2020). Weighting the domain of probability densities in functional data analysis.\emph{Stat}(2020). https://doi.org/10.1002/sta4.283
}
\author{
R. Talska\email{talskarenata@seznam.cz}, A. Menafoglio, K. Hron\email{karel.hron@upol.cz}, J. J. Egozcue, J. Palarea-Albaladejo
}
