% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transformations.R
\name{fcenLRp}
\alias{fcenLRp}
\title{fcenLRp transformation (functional)}
\usage{
fcenLRp(z, z_step, density, p)
}
\arguments{
\item{z}{grid of points defining the abscissa}

\item{z_step}{step of the grid of the abscissa}

\item{density}{grid evaluation of the P-density}

\item{p}{density of the reference measure P}
}
\value{
\item{\code{out}}{grid evaluation of the P-density in L2(P)}
}
\description{
fcenLR[P] transformation: mapping from B2(P) into L2(P)
}
\references{
Talska, R., Menafoglio, A., Hron, K., Egozcue, J. J., Palarea-Albaladejo, J. (2020). Weighting the domain of probability densities in functional data analysis.\emph{Stat}(2020). https://doi.org/10.1002/sta4.283
}
\author{
R. Talska\email{talskarenata@seznam.cz}, A. Menafoglio, K. Hron\email{karel.hron@upol.cz}, J.J. Egozcue, J. Palarea-Albaladejo
}
