% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSets.R
\docType{data}
\name{GDPsatis}
\alias{GDPsatis}
\title{GDP satisfaction}
\format{
A data frame with 31 observations and 8 variables
}
\source{
from Eurostat,\url{https://ec.europa.eu/eurostat/}
}
\usage{
data(GDPsatis)
}
\description{
Satisfaction of GDP in 31 countries. The GDP is measured per capita from the year 2012.
}
\details{
\itemize{
 \item{\code{country }}{community code}
 \item{\code{gdp }}{GDP per capita in 2012}
 \item{\code{very.bad }}{satisfaction very bad}
 \item{\code{bad }}{satisfaction bad}
 \item{\code{moderately.bad }}{satisfaction moderately bad}
 \item{\code{moderately.good }}{satisfaction moderately good}
 \item{\code{good }}{satisfaction good}
 \item{\code{very.good }}{satisfaction very good}
}
}
\examples{

data(GDPsatis)
str(GDPsatis)
}
\author{
Peter Filzmoser, Matthias Templ
}
\keyword{data}
