% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gm}
\alias{gm}
\title{gmean}
\usage{
gm(data)
}
\arguments{
\item{data}{a vector}
}
\description{
This function calculates the geometric mean.
}
\details{
\code{gm} calculates the geometric mean for all positive entries of a vector. It is implemented
in C++ and accounts for over- and underflows.
}
\examples{
gm(c(3,5,3,6,7))
}
\author{
Matthias Templ
}

