% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSets.R
\docType{data}
\name{honey}
\alias{honey}
\title{honey compositions}
\format{
A data frame with 429 observations on the following 17 variables.
\itemize{ 
\item{\code{class }}{adulterated honey, Honey or Syrup}
\item{\code{group }}{group information}
\item{\code{group3 }}{detailed group information}
\item{\code{group1 }}{less detailed group information}
\item{\code{region }}{region}
\item{\code{Al }}{chemical element}
\item{\code{B }}{chemical element}
\item{\code{Ba }}{chemical element} 
\item{\code{Ca }}{chemical element} 
\item{\code{Fe }}{chemical element}
\item{\code{K }}{chemical element}
\item{\code{Mg }}{chemical element} 
\item{\code{Mn}}{chemical element} 
\item{\code{Na }}{chemical element} 
\item{\code{P }}{chemical element} 
\item{\code{Sr }}{chemical element} 
\item{\code{Zn }}{chemical element} 
}
}
\source{
Mendeley Data, contributed by Liping Luo and translated to R by Matthias Templ
}
\description{
The contents of honey, syrup, and adulteration mineral elements.
}
\details{
Discrimination of honey and adulteration by elemental chemometrics profiling.
}
\note{
In the original paper, sparse PLS-DA were applied optimize the classify model 
and test effectiveness. Classify accuracy were exceed 87.7 percent.
}
\examples{

data(honey)

}
\references{
Tao Liu, Kang Ming, Wei Wang, Ning Qiao, Shengrong Qiu, Shengxiang Yi, Xueyong Huang, Liping Luo,
Discrimination of honey and syrup-based adulteration by mineral element chemometrics profiling,'
Food Chemistry, Volume 343, 2021, \doi{10.1016/j.foodchem.2020.128455}.
}
\keyword{datasets}
