% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ilr.2x2.R
\name{ilr.2x2}
\alias{ilr.2x2}
\title{ilr coordinates in 2x2 tables}
\usage{
ilr.2x2(x)
}
\arguments{
\item{x}{a 2x2 table}
}
\value{
The ilr coordinates
}
\description{
ilr coordinates of original, independent and interaction table using SBP1 and SBP2
}
\examples{
data(employment) 
ilr.2x2(employment[,,"AUT"])
ilr.2x2(employment)
}
\author{
Kamila Facevicova, Matthias Templ
}
\references{
K. Facevicova, K. Hron, V. Todorov, D. Guo, M. Templ (2014).
Logratio approach to statistical analysis of 2x2 compositional tables.
\emph{Journal of Applied Statistics}, Volume 41 (5), 944--958.
DOI:10.1080/02664763.2013.856871
}

