% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSets.R
\docType{data}
\name{instw}
\alias{instw}
\title{value added, output and input for different ISIC codes and countries.}
\format{A data.frame with 1555 rows and 7 columns.}
\usage{
data(instw)
}
\description{
\itemize{
 \item{\code{ct}}{ct}
 \item{\code{isic}}{ISIC classification, Rev 3.2}
 \item{\code{VA}}{value added}
 \item{\code{OUT}}{output}
 \item{\code{INP}}{input}
 \item{\code{IS03}}{country code}
 \item{\code{mht}}{mht}
}
}
\examples{
data(instw)
head(instw)
}
\author{
Valentin Todorov, Matthias Templ
}
\keyword{data}
