% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSets.R
\docType{data}
\name{isic32}
\alias{isic32}
\title{ISIC codes by name}
\format{A data.frame with 24 rows and 2 columns.}
\usage{
data(isic32)
}
\description{
\itemize{
 \item{\code{code}}{ISIC code, Rev 3.2}
 \item{\code{description}}{Description of ISIC codes}
}
}
\examples{
data(instw)
instw
}
\author{
Valentin Todorov, Matthias Templ
}
\keyword{data}
