% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSets.R
\docType{data}
\name{laborForce}
\alias{laborForce}
\title{labour force by status in employment}
\format{
A data set on 124 compositions on 9 variables.
}
\source{
from UNSTATS website
}
\description{
Labour force by status in employment for 124 countries, latest update: December 2009
}
\details{
\itemize{
\item{\code{country }}{country}                            
\item{\code{year }}{year}                                      
\item{\code{employeesW }}{percentage female employees}                             
\item{\code{employeesM }}{percentage male employees}
\item{\code{employersW }}{percentage female employers}                    
\item{\code{employersM }}{percentage male employers}
\item{\code{ownW }}{percentage female own-account workers and contributing family workers}               
\item{\code{ownM }}{percentage male own-account workers and contributing family workers} 
\item{\code{source }}{HS: household or labour force survey. OE: official estimates. PC: population census} 
}
}
\examples{

data(laborForce)
str(laborForce)

}
\references{
K. Hron, P. Filzmoser, K. Thompson (2012). Linear regression with compositional explanatory variables. \emph{Journal of Applied Statistics}, Volume 39, Issue 5, 2012.
}
\author{
conversion to R by Karel Hron and Matthias Templ \email{matthias.templ@tuwien.ac.at}
}
\keyword{datasets}
