% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSets.R
\docType{data}
\name{landcover}
\alias{landcover}
\title{European land cover}
\format{
A data set on 28 compositions on 7 variables.
}
\source{
Lucas land cover
}
\description{
Land cover data from Eurostat (2015) extended with (log) population and (log) pollution
}
\details{
\itemize{
\item{\code{Woodland }}{Coverage in km2}                            
\item{\code{Cropland }}{Coverage in km2}                                      
\item{\code{Grassland }}{Coverage in km2}                             
\item{\code{Water }}{Coverage in km2}
\item{\code{Artificial }}{Coverage in km2}                    
\item{\code{Pollution }}{log(Pollution) values per country}
\item{\code{PopDensity }}{log(PopDensity) values per country}               
}
}
\examples{

data(landcover)
str(landcover)

}
\author{
conversion to R by Karel Hron
}
\keyword{datasets}
