% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSets.R
\docType{data}
\name{lifeExpGdp}
\alias{lifeExpGdp}
\title{Life expectancy and GDP (2008) for EU-countries}
\format{A data set on 27 compositions on 9 variables.}
\source{
\url{http://www.ec.europa.eu/eurostat} and \url{http://unstats.un.org/unsd}
}
\description{
Social-economic data for compositional regression.
}
\details{
\itemize{
\item{\code{country }}{country}                                   
\item{\code{agriculture }}{GDP on agriculture, hunting, forestry, fishing (ISIC A-B, x1)}               
\item{\code{manufacture }}{GDP on mining, manufacturing, utilities (ISIC C-E, x2)}
\item{\code{construction }}{GDP on construction (ISIC F, x3)}              
\item{\code{wholesales }}{GDP on wholesale, retail trade, restaurants and hotels (ISIC G-H, x4)}
\item{\code{transport }}{GDP on transport, storage and communication (ISIC I, x5)}
\item{\code{other }}{GDP on other activities (ISIC J-P, x6)}
\item{\code{lifeExpMen }}{life expectancy for men and women}
\item{\code{lifeExpWomen }}{life expectancy for men and women}
}
}
\examples{

data(lifeExpGdp)
str(lifeExpGdp)
}
\author{
conversion to R by Karel Hron and Matthias Templ \email{matthias.templ@tuwien.ac.at}
}
\references{
K. Hron, P. Filzmoser, K. Thompson (2012). Linear regression with compositional explanatory variables. \emph{Journal of Applied Statistics}, Volume 39, Issue 5, 2012.
}
\keyword{datasets}

