% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dataSets.R
\docType{data}
\name{mcadd}
\alias{mcadd}
\title{Metabolomics MCADD data set}
\format{A data frame with 50 observations and 279 variables}
\usage{
data(mcadd)
}
\description{
\itemize{
 \item{\code{group }}{patient group}
 \item{\code{... }}{the remaining variables named after the time stamps.}
}
}
\examples{
data(mcadd)
str(mcadd)
}
\keyword{data}

