% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSets.R
\docType{data}
\name{mortality_tab}
\alias{mortality_tab}
\title{mortality table}
\format{
A table
}
\usage{
data(mortality_tab)
}
\description{
Mortality data by gender, unknown year
}
\details{
\itemize{
 \item{\code{female}}{mortality rates for females by age groups}
 \item{\code{male}}{mortality rates for males by age groups}
}
}
\examples{
data(mortality_tab)
mortality_tab
}
\author{
Matthias Templ
}
\keyword{data}
