% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm1.R
\name{norm1}
\alias{norm1}
\title{Normalize a vector to length 1}
\usage{
norm1(x)
}
\arguments{
\item{x}{a numeric vector}
}
\description{
Scales a vector to a unit vector.
}
\examples{

data(expenditures)
i <- 1
D <- 6
vec <- c(rep(-1/i, i), 1, rep(0, (D-i-1)))

norm1(vec)

}
\author{
Matthias Templ
}
\keyword{manip}
