% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSets.R
\docType{data}
\name{nutrients_branded}
\alias{nutrients_branded}
\title{nutrient contents (branded)}
\format{A data frame with 9618 observations on the following 18 variables.}
\source{
From the Swiss nutrition data base 2015 (second edition), see \url{ http://www.sge-ssn.ch/shop/produkt/schweizer-naehrwerttabelle/}
}
\usage{
data(nutrients_branded)
}
\description{
Nutrients on more than 10 components and 9618 branded food products
}
\details{
\itemize{
\item{\code{name_D }}{name (in German)}
\item{\code{category_D }}{factor specifying the category names}
\item{\code{category_F }}{factor specifying the category names}
\item{\code{category_I }}{factor specifying the category names}
\item{\code{category_E }}{factor specifying the category names}
\item{\code{gravity }}{specific gravity}
\item{\code{energy_kJ }}{energy in kJ}
\item{\samp{energy_kcal }}{energy in kcal}
\item{\code{protein }}{protein in gram}
\item{\code{alcohol }}{alcohol in gram}
\item{\code{water }}{water in gram}
\item{\code{carbohydrates_available }}{available carbohydrates in gram}
\item{\code{sugars }}{sugars in gram}
\item{\code{dietary_fibres }}{dietary fibres in gram}
\item{\code{fat_total }}{total fat in gram}
\item{\code{fatty_acids_saturated }}{saturated acids fat in gram}
\item{\code{Na }}{Sodium in gram}
\item{\code{unit }}{a factor with levels \code{per 100g edible portion} \code{per 100ml food volume}}
}
}
\examples{

data(nutrients_branded)
str(nutrients_branded)
}
\author{
Translated from the Swiss nutrion data base by Matthias Templ \email{matthias.templ@tuwien.ac.at}
}
\references{
\url{ http://www.sge-ssn.ch/shop/produkt/schweizer-naehrwerttabelle/}
}
\keyword{data}

