% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSets.R
\docType{data}
\name{payments}
\alias{payments}
\title{special payments}
\format{A data frame with 535 rows and 11 variables}
\source{
statCube data base at \url{www.statistik.ac.at}. The product and all 
material contained therein are protected by copyright with all rights 
reserved by the Bundesanstalt Statistik Oesterreich (STATISTICS AUSTRIA). 
It is permitted to reproduce, distribute, make publicly available 
and process the content for non-commercial purposes. Prior to any use for 
commercial purposes a written consent of STATISTICS AUSTRIA must be obtained. 
Any use of the contained material must 
be correctly reproduced and clearly cite the source STATISTICS AUSTRIA. 
If tables published by STATISTICS AUSTRIA are partially used, displayed or 
otherwise changed, a note must be added at an adequate position to 
show data was extracted or adapted.
}
\usage{
data(payments)
}
\description{
Payments splitted by different NACE categories and kind of employment in Austria 2004
}
\details{
\itemize{
 \item{\code{nace }}{NACE classification, 2 digits}
 \item{\code{oenace_2008 }}{Corresponding Austrian NACE classification (in German)}
 \item{\code{year }}{year}
 \item{\code{month }}{month}
 \item{\code{localunit }}{local unit ID}
 \item{\code{spay }}{special payments (total)}
 \item{\code{spay_wc }}{special payments for white colar workers}
 \item{\code{spay_bc }}{special payments for blue colar workers}
 \item{\code{spay_traintrade }}{special payments for trainees in trade businness}
 \item{\code{spay_home }}{special payments for home workers}
 \item{\code{spay_traincomm }}{special payments for trainees in commercial businness}
}
}
\examples{

data(payments)
str(payments)
summary(payments)
}
\author{
Matthias Templ \email{matthias.templ@tuwien.ac.at}
}
\keyword{data}
