% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perturbation.R
\name{perturbation}
\alias{perturbation}
\alias{powering}
\title{Perturbation and powering}
\usage{
perturbation(x, y)

powering(x, a)
}
\arguments{
\item{x}{(compositional) vector containing positive values}

\item{y}{(compositional) vector containing positive values or NULL for powering}

\item{a}{constant, numeric vector of length 1}
}
\value{
Result of perturbation or powering
}
\description{
Perturbation and powering for two compositions.
}
\examples{
data(expenditures)
x <- expenditures[1 ,]
y <- expenditures[2, ]
perturbation(x, y)
powering(x, 2)
}
\references{
Aitchison, J. (1986) \emph{The Statistical Analysis of
Compositional Data} Monographs on Statistics and Applied Probability.
Chapman and Hall Ltd., London (UK). 416p.
}
\author{
Matthias Templ
}
\keyword{arith}
\keyword{math}
