% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSets.R
\docType{data}
\name{phd_totals}
\alias{phd_totals}
\title{PhD students in the EU (totals)}
\format{A data set on 29 compositions and 5 variables.}
\source{
Eurostat
}
\description{
PhD students in Europe by different kind of studies.
}
\details{
\itemize{
\item{\code{technical }}{phd students in natural and technical sciences}
\item{\code{socio-economic-low }}{phd students in social sciences, economic sciences and law sciences}                    
\item{\code{human }}{phd students in human sciences including teaching}
\item{\code{health }}{phd students in health and life sciences}               
\item{\code{agriculture }}{phd students in agriculture} 
}
}
\examples{

data("phd_totals")
str(phd_totals)

}
\references{
Hron, K. and Templ, M. and Filzmoser, P. (2010) Imputation of missing values for compositional data using classical and robust methods. \emph{Computational Statistics and Data Analysis}, vol 54 (12), pages 3095-3107.
}
\keyword{datasets}
