% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splineDensity.plot.R
\name{plot.smoothSpl}
\alias{plot.smoothSpl}
\title{plot smoothSpl}
\usage{
\method{plot}{smoothSpl}(x, y, ..., by = 1, n = 10, index = NULL)
}
\arguments{
\item{x}{class smoothSpl object}

\item{y}{ignored}

\item{...}{further arguments passed by}

\item{by}{stepsize}

\item{n}{length of sequence to plot}

\item{index}{optinally the sequence instead of by and n}
}
\description{
plot densities of objects of class smoothSpl
}
\author{
Alessia Di Blasi, Federico Pavone, Gianluca Zeni
}
