% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSets.R
\docType{data}
\name{precipitation}
\alias{precipitation}
\title{24-hour precipitation}
\format{
A table with 4 rows and 6 columns
}
\usage{
data(precipitation)
}
\description{
table containing counts for 24-hour precipitation for season at the rain-gouge.
}
\details{
\itemize{
 \item{\code{spring}}{numeric vector on counts for different level of precipitation}
 \item{\code{summer}}{numeric vector on counts for different level of precipitation}
 \item{\code{autumn}}{numeric vector on counts for different level of precipitation}
 \item{\code{winter}}{numeric vector on counts for different level of precipitation}
}
}
\examples{
data(precipitation)
precipitation
str(precipitation)
}
\references{
Romero R, Guijarro J A, Ramis C, Alonso S (1998). A 30-years (196493) daily rainfall
data base for the Spanish Mediterranean regions: first exploratory study. 
\emph{International Journal of Climatology} 18, 541560.
}
\author{
Matthias Templ \email{matthias.templ@tuwien.ac.at}
}
\keyword{data}
