% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.imp.R
\name{print.imp}
\alias{print.imp}
\title{Print method for objects of class imp}
\usage{
\method{print}{imp}(x, ...)
}
\arguments{
\item{x}{an object of class \sQuote{imp}}

\item{\dots}{additional arguments passed trough}
}
\value{
None (invisible NULL).
}
\description{
The function returns a few information about how many missing values are
imputed and possible other information about the amount of iterations, for
example.
}
\examples{

data(expenditures)
expenditures[1,3]
expenditures[1,3] <- NA
\dontrun{
xi <- impCoda(expenditures)
xi
summary(xi)
plot(xi, which=1:2)
}

}
\author{
Matthias Templ
}
\seealso{
\code{\link{impCoda}}, \code{\link{impKNNa}}
}
\keyword{print}

