% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSets.R
\docType{data}
\name{production}
\alias{production}
\title{production splitted by nationality on enterprise level}
\format{
A data frame with 535 rows and 9 variables
}
\source{
statCube data base at the website of Statistics Austria. The product and all 
material contained therein are protected by copyright with all rights 
reserved by the Bundesanstalt Statistik Oesterreich (STATISTICS AUSTRIA). 
It is permitted to reproduce, distribute, make publicly available 
and process the content for non-commercial purposes. Prior to any use for 
commercial purposes a written consent of STATISTICS AUSTRIA must be obtained. 
Any use of the contained material must 
be correctly reproduced and clearly cite the source STATISTICS AUSTRIA. 
If tables published by STATISTICS AUSTRIA are partially used, displayed or 
otherwise changed, a note must be added at an adequate position to 
show data was extracted or adapted.
}
\usage{
data(production)
}
\description{
\itemize{
 \item{\code{nace }}{NACE classification, 2 digits}
 \item{\code{oenace_2008 }}{Corresponding Austrian NACE classification (in German)}
 \item{\code{year }}{year}
 \item{\code{month }}{month}
 \item{\code{enterprise }}{enterprise ID}
 \item{\code{total }}{total ...}
 \item{\code{home }}{home ...}
 \item{\code{EU }}{EU ...}
 \item{\code{non-EU }}{non-EU ...}
}
}
\examples{

data(production)
str(production)
summary(production)
}
\author{
Matthias Templ \email{matthias.templ@tuwien.ac.at}
}
\keyword{data}
