\name{robGUI}
\alias{robGUI}
\title{
A lightweighted graphical user interface
}
\description{
A lightweighted graphical user interface for the robCompositions package written in Gkt2. Before calling the GUI, RGtk2 have to be installed.
}
\usage{
robGUI()
}
\details{
Already some of the methods from the robCompositions package can be applied by this GUI. The application is straighforward and worth to mention.
}
\value{
The GUI.
}
\author{
Jiri Eichler
}
\note{
The next version of the GUI may include tools for reproducibility (save and load/run script files).
}
\examples{
## Not run:
## to open the GUI:
## robGUI()
## End(Not run)
}
\keyword{ utilities }
