% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSDev.R
\name{rSDev}
\alias{rSDev}
\title{Relative simplicial deviance}
\usage{
rSDev(x, y)
}
\arguments{
\item{x}{a propability table}

\item{y}{an interaction table}
}
\value{
The relative simplicial deviance
}
\description{
Relative simplicial deviance
}
\examples{
data(precipitation) 
tabprob <- prop.table(precipitation)
tabind <- indTab(precipitation)
tabint <- intTab(tabprob, tabind)
rSDev(tabprob, tabint$intTab)
}
\references{
Egozcue, J.J., Pawlowsky-Glahn, V., Templ, M., Hron, K. (2015)
Independence in contingency tables using simplicial geometry. 
\emph{Communications in Statistics - Theory and Methods}, 44 (18), 3978--3996.
}
\author{
Matthias Templ
}
