% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSets.R
\docType{data}
\name{saffron}
\alias{saffron}
\title{saffron compositions}
\format{
A data frame with 53 observations on the following 36 variables.
\itemize{ 
\item{\code{Sample }}{adulterated honey, Honey or Syrup}
\item{\code{Country }}{group information}
\item{\code{Batch }}{detailed group information}
\item{\code{Region }}{less detailed group information}
\item{\code{d2H }}{region}
\item{\code{d13C }}{chemical element}
\item{\code{d15N }}{chemical element}
\item{\code{Li }}{chemical element} 
\item{\code{B }}{chemical element} 
\item{\code{Na }}{chemical element}
\item{\code{Mg }}{chemical element}
\item{\code{Al }}{chemical element} 
\item{\code{K}}{chemical element} 
\item{\code{Ca }}{chemical element} 
\item{\code{V }}{chemical element} 
\item{\code{Mn }}{chemical element} 
\item{\code{Fe }}{chemical element} 
\item{\code{Co }}{chemical element} 
\item{\code{Ni }}{chemical element} 
\item{\code{Cu }}{chemical element} 
\item{\code{Zn }}{chemical element} 
\item{\code{Ga }}{chemical element} 
\item{\code{As }}{chemical element} 
\item{\code{Rb }}{chemical element} 
\item{\code{Sr }}{chemical element} 
\item{\code{Y }}{chemical element} 
\item{\code{Mo }}{chemical element} 
\item{\code{Cd }}{chemical element} 
\item{\code{Cs }}{chemical element} 
\item{\code{Ba }}{chemical element} 
\item{\code{Ce }}{chemical element} 
\item{\code{Pr }}{chemical element} 
\item{\code{Nd }}{chemical element} 
\item{\code{Sm }}{chemical element} 
\item{\code{Gd }}{chemical element} 
\item{\code{Pb }}{chemical element} 
}
}
\source{
Mendeley Data, contributed by Russell Frew and translated to R by Matthias Templ
}
\description{
Stable isotope ratio  and trace metal cncentration data for saffron samples.
}
\note{
In the original paper, the authors applied lda for classifying the observations.
}
\examples{

data(saffron)

}
\references{
Frew, Russell (2019), Data for: 
CHEMICAL PROFILING OF SAFFRON FOR AUTHENTICATION OF ORIGIN, Mendeley Data, V1, 
\doi{10.17632/5544tn9v6c.1}
}
\keyword{datasets}
