% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDev.R
\name{SDev}
\alias{SDev}
\title{Simplicial deviance}
\usage{
SDev(x)
}
\arguments{
\item{x}{a propability table}
}
\value{
The simplicial deviance
}
\description{
Simplicial deviance
}
\examples{
data(precipitation) 
tab1prob <- prop.table(precipitation)
SDev(tab1prob)
}
\references{
Juan Jose Egozcuea, Vera Pawlowsky-Glahn, Matthias Templ, Karel Hron (2015)
Independence in Contingency Tables Using Simplicial Geometry. 
\emph{Communications in Statistics - Theory and Methods}, Vol. 44 (18), 3978--3996.
DOI:10.1080/03610926.2013.824980
}
\author{
Matthias Templ
}
