\name{skyeLavas}
\alias{skyeLavas}
\docType{data}
\title{ Aphyric skye lavas data }
\description{
  AFM compositions of 23 aphyric Skye lavas. This data set can be found on page 360 of the Aitchison book (see reference).
}
\usage{data(skyeLavas)}
\format{
  A data frame with 23 observations on the following 3 variables.
  \describe{
    \item{\code{sodium-potassium}}{a numeric vector of percentages of Na2O\eqn{+}K2O}
    \item{\code{iron}}{a numeric vector of percentages of Fe2O3}
    \item{\code{magnesium}}{a numeric vector of percentages of MgO}
  }
}
\source{
Aitchison, J. (1986) \emph{The Statistical Analysis of Compositional
Data} Monographs on Statistics and Applied Probability. Chapman \&
Hall Ltd., London (UK). 416p.
}
\examples{
data(skyeLavas)
}
\keyword{datasets}
