% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.imp.R
\name{summary.imp}
\alias{summary.imp}
\title{Summary method for objects of class imp}
\usage{
\method{summary}{imp}(object, ...)
}
\arguments{
\item{object}{an object of class \sQuote{imp}}

\item{\dots}{additional arguments passed trough}
}
\value{
None (invisible NULL).
}
\description{
A short comparison of the original data and the imputed data is given.
}
\details{
Note that this function will be enhanced with more sophisticated methods in
future versions of the package.  It is very rudimental in its present form.
}
\examples{

data(expenditures)
expenditures[1,3]
expenditures[1,3] <- NA
xi <- impKNNa(expenditures)
xi
summary(xi)
# plot(xi, which=1:2)

}
\seealso{
\code{\link{impCoda}}, \code{\link{impKNNa}}
}
\author{
Matthias Templ
}
\keyword{print}
