% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transformations.R
\name{trapzc}
\alias{trapzc}
\title{Trapezoidal formula for numerical integration}
\usage{
trapzc(step, f)
}
\arguments{
\item{step}{step of the grid}

\item{f}{density from fcenLR}
}
\value{
\item{\code{int}}{The value of integral computed numerically by trapezoidal formula.}
}
\description{
Numerical integration via trapezoidal formula.
}
\examples{
# Example (zero-integral of fcenLR density)
t = seq(-4.7,4.7, length = 1000)
t_step = diff(t[1:2])
mean = 0; sd = 1.5
dens.true = dnorm(t, mean, sd)
dens.true.l = fcenLR(t,t_step,dens.true)
trapzc(t_step,dens.true.l)
}
\author{
R. Talska\email{talskarenata@seznam.cz}, K. Hron\email{karel.hron@upol.cz}
}
