% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightedPivortCoord.R
\name{weightedPivotCoord}
\alias{weightedPivotCoord}
\title{Weighted pivot coordinates}
\usage{
weightedPivotCoord(
  x,
  pivotvar = 1,
  option = "var",
  method = "classical",
  pow = 1,
  yvar = NULL
)
}
\arguments{
\item{x}{object of class `data.frame` or `matrix`; positive values only}

\item{pivotvar}{pivotal variable; if any other number than 1, the data are resorted in that sense that 
pivotvar is shifted to the first part}

\item{option}{option for the choice of weights. If `option = "var"` (default), weights are based on variation matrix elements: 
`(1/t_{1j})^pow`, if `option = "cor"`, weights are based on correlations between variable specified in yvar and logratios and its distribution:
`|integral_{0}^{r_{j}} f(x) dx|`, `f(x)...` Kernel density estimator for `s_{j}; s_{j}=0 if |r_{j}|<cut` otherwise `s_{j}=r_{j}`,
`cut = min(#r_{j}=>0/#r_{j}, #r_{j}<0/#r_{j}`, with Gaussian Kernel function and bandwidth `h=0.05`.}

\item{method}{method for estimation of variation/correlation,
if `option = "classical"` (default), classical estimation is applied,
if `option = "robust"`, robust estimation is applied;}

\item{pow}{if `option = "var"`, power `pow` is applied on unnormalized weights; default is 1;}

\item{yvar}{if `option = "cor"`, weights are based on correlation between logratios and variable specified in `yvar`;}
}
\value{
\item{WPC}{weighted pivot coordinates (matrix with n rows and (D-1) columns)}
\item{w}{logcontrasts (matrix with D rows and (D-1) columns)}
}
\description{
XXX
}
\details{
weightedPivotCoord

XXX
}
\examples{
###################
### XXX

data(expenditures)
w <- weightedPivotCoord(expenditures)
w
}
\references{
XXX
}
\seealso{
\code{\link{pivotCoord}}
}
\author{
Nikola.Stefelova
}
\keyword{multivariate}
