% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeroOut.R
\name{zeroOut}
\alias{zeroOut}
\title{Detection of outliers of zero-inflated data}
\usage{
zeroOut(x, impute = "knn")
}
\arguments{
\item{x}{a data frame}

\item{impute}{imputation method internally used}
}
\value{
XXX
}
\description{
detects outliers in compositional zero-inflated data
}
\details{
XXX
}
\examples{
### Installing and loading required packages
data(expenditures)
}
\author{
Matthias Templ
}

