% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSets.R
\docType{data}
\name{alcoholreg}
\alias{alcoholreg}
\title{regional alcohol per capita (15+) consumption by WHO region}
\format{
A data frame with 6 rows and 4 variables
}
\source{
Transfered from the World Health Organisation website.
}
\usage{
data(alcoholreg)
}
\description{
\itemize{
 \item{\code{country }}{Country}
 \item{\code{year }}{Year}
 \item{\code{recorded }}{Recorded alcohol consumption}
 \item{\code{unrecorded }}{Unrecorded alcohol consumption}
}
}
\examples{

data("alcoholreg")
alcoholreg
}
\author{
Matthias Templ \email{matthias.templ@tuwien.ac.at}
}
\keyword{data}
