% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeBDLs.R
\name{bootnComp}
\alias{bootnComp}
\title{Bootstrap to find optimal number of components}
\usage{
bootnComp(X, y, R = 99, plotting = FALSE)
}
\arguments{
\item{X}{predictors as a matrix}

\item{y}{response}

\item{R}{number of bootstrap replicates}

\item{plotting}{if TRUE, a diagnostic plot is drawn for each bootstrap
replicate}
}
\value{
Including other information in a list, the optimal number of
components
}
\description{
Combined bootstrap and cross validation procedure to find optimal number of
PLS components
}
\details{
Heavily used internally in function impRZilr.
}
\examples{

## we refer to impRZilr()

}
\seealso{
\code{\link{impRZilr}}
}
\author{
Matthias Templ
}
\keyword{manip}
