% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSets.R
\docType{data}
\name{coffee}
\alias{coffee}
\title{coffee data set}
\format{
A data frame with 30 observations and 7 variables.
}
\usage{
data(coffee)
}
\description{
30 commercially available coffee samples of different origins.
}
\details{
\itemize{
\item{\code{sort }}{sort of coffee}
\item{\code{acit }}{acetic acid }
\item{\code{metpyr }}{methylpyrazine}
\item{\code{furfu }}{furfural}
\item{\code{furfualc }}{furfuryl alcohol}
\item{\code{dimeth }}{2,6 dimethylpyrazine}
\item{\code{met5 }}{5-methylfurfural}
}

In the original data set, 15 volatile compounds (descriptors of coffee aroma) were selected for a statistical analysis. We selected six compounds (compositional parts) on three sorts of coffee.
}
\examples{

data(coffee)
str(coffee)
summary(coffee)
}
\references{
M. Korhonov\'a, K. Hron, D. Klimc\'ikov\'a, L. Muller, P. Bedn\'ar, and P. Bart\'ak (2009). Coffee aroma - statistical analysis of compositional data. \emph{Talanta}, 80(2): 710--715.
}
\author{
Matthias Templ \email{matthias.templ@tuwien.ac.at}, Karel Hron
}
\keyword{data}
