% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSets.R
\docType{data}
\name{employment}
\alias{employment}
\title{employment in different countries by gender and status.}
\format{A three-dimensional table}
\usage{
data(employment)
}
\description{
employment in different countries by gender and status.
}
\examples{
data(employment)
str(employment)
employment
}
\keyword{data}
