% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSets.R
\docType{data}
\name{foodbalance}
\alias{foodbalance}
\title{country food balances}
\format{
A data frame with 115 observations on the following 116 variables.
\itemize{ 
\item{\code{country}}{country} 
\item{\code{Cereals - Excluding Beer}}{Food balance on cereals} 
\item{\code{...}}{...}
#' \item{\code{Alcohol - Non-Food}}{Food balance on alcohol}
}
}
\source{
\url{http://www.fao.org/home/en/}
}
\description{
Food balance in each country (2018)
}
\examples{

data(foodbalance)

}
\keyword{datasets}
