% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm.R
\name{gm}
\alias{gm}
\title{geometric mean}
\usage{
gm(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
The geometric mean.
}
\description{
This function calculates the geometric mean.
}
\details{
Calculates the geometric mean of all positive entries of a vector.
}
\examples{

gm(runif(100))

}
\author{
Matthias Templ
}
\keyword{math}

