% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSets.R
\docType{data}
\name{haplogroups}
\alias{haplogroups}
\title{Haplogroups data.}
\format{A data frame with 38 observations on the following 12 variables.
\itemize{ 
\item{\code{I1 }}{pre-Germanic (Nordic)}
\item{\code{I2b }}{pre-Celto-Germanic} 
\item{\code{I2a1 }}{Sardinian, Basque} 
\item{\code{I2a2 }}{Dinaric, Danubian}
\item{\code{N1c1 }}{Uralo-Finnic, Baltic, Siberian}
\item{\code{R1a }}{Balto-Slavic, Mycenaean Greek, Macedonia}
\item{\code{R1b }}{Italic, Celtic, Germanic; Hitite, Armenian}
\item{\code{G2a }}{Caucasian, Greco-Anatolien} 
\item{\code{E1b1b }}{North and Eastern Afrika, Near Eastern, Balkanic} 
\item{\code{J2 }}{Mesopotamian, Minoan Greek, Phoenician} 
\item{\code{J1 }}{Semitic (Arabic, Jewish)}
\item{\code{T }}{Near-Eastern, Egyptian, Ethiopian, Arabic} 
}}
\source{
Eupedia:
\url{http://www.eupedia.com/europe/european_y-dna_haplogroups.shtml}
}
\description{
Distribution of European Y-chromosome DNA (Y-DNA) haplogroups by region in
percentage.
}
\details{
Human Y-chromosome DNA can be divided in genealogical groups sharing a
common ancestor, called haplogroups.
}
\examples{

data(haplogroups)

}
\keyword{datasets}

