% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSets.R
\docType{data}
\name{machineOperators}
\alias{machineOperators}
\title{machine operators}
\format{
A data frame with 27 observations on the following 4 variables.
}
\usage{
data(machineOperators)
}
\description{
Compositions of eight-hour shifts of 27 machine operators
}
\details{
\itemize{
\item{\code{hqproduction }}{high-quality production}
\item{\code{lqproduction }}{low-quality production}
\item{\code{setting }}{machine settings}
\item{\code{repair }}{machine repair}
}

The data set from Aitchison (1986), p. 382, contains compositions of eight-hour shifts of 27 machine operators. The parts represent proportions of shifts in each activity:  high-quality production, low-quality production, machine setting and machine repair.
}
\examples{

data(machineOperators)
str(machineOperators)
summary(machineOperators)
rowSums(machineOperators)
}
\references{
Aitchison, J. (1986) \emph{The Statistical Analysis of Compositional Data} Monographs on Statistics and Applied Probability. Chapman and Hall Ltd., London (UK). 416p.
}
\author{
Matthias Templ \email{matthias.templ@tuwien.ac.at}
}
\keyword{data}
