% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.pcaCoDa.R
\name{plot.pcaCoDa}
\alias{plot.pcaCoDa}
\title{Plot method}
\usage{
\method{plot}{pcaCoDa}(x, y, ...)
}
\arguments{
\item{x}{object of class \sQuote{pcaCoDa}}

\item{y}{...}

\item{\dots}{...}
}
\value{
The robust compositional screeplot.
}
\description{
Provides a screeplot for (robust) compositional principal components analysis.
}
\examples{

data(coffee)
p1 <- pcaCoDa(coffee[,-1])
plot(p1)
plot(p1, type="lines")


}
\references{
Filzmoser, P., Hron, K., Reimann, C. (2009) Principal Component Analysis for
Compositional Data with Outliers. \emph{Environmetrics}, \bold{20} (6),
621--632.
}
\seealso{
\code{\link{pcaCoDa}}, \code{\link{biplot.pcaCoDa}}
}
\author{
M. Templ, K. Hron
}
\keyword{aplot}
