% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSets.R
\docType{data}
\name{rcodes}
\alias{rcodes}
\title{codes for UNIDO tables}
\format{
A data.frame with 2717 rows and 6 columns.
}
\usage{
data(rcodes)
}
\description{
\itemize{
 \item{\code{ISOCN}}{ISOCN codes}
 \item{\code{OPERATOR}}{Operator}
 \item{\code{ADESC}}{Country}
 \item{\code{CCODE}}{Country code}
 \item{\code{CDESC}}{Country destination}
 \item{\code{ACODE}}{Country destination code}
}
}
\examples{
data(rcodes)
str(rcodes)
}
\keyword{data}
