% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSets.R
\docType{data}
\name{skyeLavas}
\alias{skyeLavas}
\title{aphyric skye lavas data}
\format{
A data frame with 23 observations on the following 3 variables.
}
\usage{
data(skyeLavas)
}
\description{
AFM compositions of 23 aphyric Skye lavas. This data set can be found on page 360 of the Aitchison book (see reference).
}
\details{
\itemize{
\item{\code{sodium-potassium }}{a numeric vector of percentages of Na2O\eqn{+}K2O}
\item{\code{iron }}{a numeric vector of percentages of Fe2O3}
\item{\code{magnesium }}{a numeric vector of percentages of MgO}
}
}
\examples{

data(skyeLavas)
str(skyeLavas)
summary(skyeLavas)
rowSums(skyeLavas)
}
\references{
Aitchison, J. (1986) \emph{The Statistical Analysis of Compositional Data} Monographs on Statistics and Applied Probability. Chapman \& Hall Ltd., London (UK). 416p.
}
\author{
Matthias Templ \email{matthias.templ@tuwien.ac.at}
}
\keyword{data}
