% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSets.R
\docType{data}
\name{teachingStuff}
\alias{teachingStuff}
\title{teaching stuff}
\format{
A (tidy) data frame with 1216 observations on the following 4 variables.
\itemize{ 
\item{\code{country }}{Country of origin}
\item{\code{subject }}{school type: primary, lower secondary, higher secondary and tertiary } 
\item{\code{year }}{Year} 
\item{\code{value }}{Number of stuff}
}
}
\source{
OECD:
\url{https://data.oecd.org/}
}
\description{
Teaching stuff in selected countries
}
\details{
Teaching staff include professional personnel directly 
involved in teaching students, including classroom 
teachers, special education teachers and other 
teachers who work with students as a whole class, 
in small groups, or in one-to-one teaching. 
Teaching staff also include department chairs 
of whose duties include some teaching, but 
it does not include non-professional personnel 
who support teachers in providing instruction 
to students, such as teachers' aides and other 
paraprofessional personnel. Academic staff include 
personnel whose primary assignment is instruction, 
research or public service, holding an academic 
rank with such titles as professor, associate 
professor, assistant professor, instructor, 
lecturer, or the equivalent of any of these 
academic ranks. The category includes personnel 
with other titles (e.g. dean, director, associate 
dean, assistant dean, chair or head of department), 
if their principal activity is instruction or research.
}
\examples{

data(teachingStuff)
str(teachingStuff)
}
\references{
OECD (2017), Teaching staff (indicator). doi: 10.1787/6a32426b-en (Accessed on 27 March 2017)
}
\author{
translated from \url{https://data.oecd.org/} and restructured by Matthias Templ
}
\keyword{datasets}
